# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Customer Intake - Cartridge & Ink Shop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
  details?: string;
}

export interface DeleteResponse {
  success: boolean;
  message: string;
}

export interface IntakeRecord {
  id: number;
  title?: string | null;
  firstName: string;
  lastName: string;
  mobileNumber: string;
  email?: string | null;
  customerType?: string | null;
  address?: string | null;
  printerMake?: string | null;
  printerModel?: string | null;
  cartridgeTonerType?: string | null;
  serviceType: string;
  deviceSerialNumber?: string | null;
  paymentMethod?: string | null;
  notes?: string | null;
  serviceDate: string;
  serviceTime: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateIntakeRecord {
  title?: string | null;
  firstName: string;
  lastName: string;
  mobileNumber: string;
  email?: string | null;
  customerType?: string | null;
  address?: string | null;
  printerMake?: string | null;
  printerModel?: string | null;
  cartridgeTonerType?: string | null;
  serviceType: string;
  deviceSerialNumber?: string | null;
  paymentMethod?: string | null;
  notes?: string | null;
  serviceDate: string;
  serviceTime: string;
}

export type AnalyticsSummaryServiceTypeBreakdownItem = {
  serviceType: string;
  count: number;
};

export interface AnalyticsSummary {
  todayCount: number;
  weekCount: number;
  monthCount: number;
  totalCount: number;
  serviceTypeBreakdown: AnalyticsSummaryServiceTypeBreakdownItem[];
  mostCommonService?: string | null;
}

export type GetRecordsParams = {
  search?: string;
  mobile?: string;
  serviceType?: string;
  printerModel?: string;
  dateFrom?: string;
  dateTo?: string;
  sortBy?: GetRecordsSortBy;
};

export type GetRecordsSortBy =
  (typeof GetRecordsSortBy)[keyof typeof GetRecordsSortBy];

export const GetRecordsSortBy = {
  latest: "latest",
  oldest: "oldest",
  serviceDate: "serviceDate",
} as const;

export type LookupByMobileParams = {
  mobile: string;
};