# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Customer Intake - Cartridge & Ink Shop API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: records
    description: Customer intake records
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /records:
    get:
      operationId: getRecords
      tags: [records]
      summary: Get all records
      description: Returns all customer intake records, sorted latest first
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: mobile
          in: query
          schema:
            type: string
        - name: serviceType
          in: query
          schema:
            type: string
        - name: printerModel
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [latest, oldest, serviceDate]
      responses:
        "200":
          description: List of records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IntakeRecord"

    post:
      operationId: createRecord
      tags: [records]
      summary: Create a new intake record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIntakeRecord"
      responses:
        "201":
          description: Created record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeRecord"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /records/analytics:
    get:
      operationId: getAnalytics
      tags: [records]
      summary: Get analytics summary
      description: Returns analytics like totals, service type breakdown
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"

  /records/lookup-mobile:
    get:
      operationId: lookupByMobile
      tags: [records]
      summary: Lookup customer by mobile number
      description: Returns the most recent record for a mobile number for auto-fill
      parameters:
        - name: mobile
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /records/{id}:
    get:
      operationId: getRecord
      tags: [records]
      summary: Get a single record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    put:
      operationId: updateRecord
      tags: [records]
      summary: Update an intake record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIntakeRecord"
      responses:
        "200":
          description: Updated record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    delete:
      operationId: deleteRecord
      tags: [records]
      summary: Delete an intake record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required:
        - error

    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    IntakeRecord:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        mobileNumber:
          type: string
        email:
          type: string
          nullable: true
        customerType:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        printerMake:
          type: string
          nullable: true
        printerModel:
          type: string
          nullable: true
        cartridgeTonerType:
          type: string
          nullable: true
        serviceType:
          type: string
        deviceSerialNumber:
          type: string
          nullable: true
        paymentMethod:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        serviceDate:
          type: string
        serviceTime:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - firstName
        - lastName
        - mobileNumber
        - serviceType
        - serviceDate
        - serviceTime
        - createdAt
        - updatedAt

    CreateIntakeRecord:
      type: object
      properties:
        title:
          type: string
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        mobileNumber:
          type: string
        email:
          type: string
          nullable: true
        customerType:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        printerMake:
          type: string
          nullable: true
        printerModel:
          type: string
          nullable: true
        cartridgeTonerType:
          type: string
          nullable: true
        serviceType:
          type: string
        deviceSerialNumber:
          type: string
          nullable: true
        paymentMethod:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        serviceDate:
          type: string
        serviceTime:
          type: string
      required:
        - firstName
        - lastName
        - mobileNumber
        - serviceType
        - serviceDate
        - serviceTime

    AnalyticsSummary:
      type: object
      properties:
        todayCount:
          type: integer
        weekCount:
          type: integer
        monthCount:
          type: integer
        totalCount:
          type: integer
        serviceTypeBreakdown:
          type: array
          items:
            type: object
            properties:
              serviceType:
                type: string
              count:
                type: integer
            required:
              - serviceType
              - count
        mostCommonService:
          type: string
          nullable: true
      required:
        - todayCount
        - weekCount
        - monthCount
        - totalCount
        - serviceTypeBreakdown